+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 ## Unreleased
 - added `@Frames` macro for full power mode of defining frames
 
+## v0.5.3 (26th of July 2021)
+- Allow all kinds of iterable ways in the `act!` function such that `act!(::Matrix, ::Action)` also works
+- Updated `anim_translate`
+  - Docstring: `anim_translate` translates by a vector instead of to a point
+  - from->to assumes that we are at `from` already instead of adding it to it
+
 ## v0.5.2
 - Added support for local network live streaming